requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
umap-learn>=0.5
librosa>=0.10
soundfile>=0.12
tqdm>=4.66
torch>=2.1
datasets>=2.18
datasets[audio]>=2.18
huggingface_hub[cli]>=0.20
